Fixes incorrect path for system files
