+++ src/app_data.cpp
@@ -1,5 +1,5 @@
 #include "app_data.hpp"
-#include "util.hpp"
+#include "path.hpp"
 
 #include <QStandardPaths>
